requirements.txt
# Vector database
chromadb>=0.4.22

# Embeddings
sentence-transformers>=2.2.2

# Document parsing
pypdf>=3.17.0
python-docx>=1.1.0
pytesseract>=0.3.10
Pillow>=10.0.0

# Text processing
langchain-text-splitters>=0.0.1

# LLM providers
openai>=1.12.0
ollama>=0.1.6

# Utilities
python-dotenv>=1.0.0

# Web UI
gradio>=5.0.0

# Testing
pytest>=7.4.0
